Update download logic and fix README typos.
