+++ demos/speech_translation/README.md
@@ -22,7 +22,7 @@ wget https://paddlespeech.bj.bcebos.com/PaddleAudio/zh.wav https://paddlespeech.
 ### 3. Usage
 - Command Line(Recommended)
   ```bash
-  paddlespeech st --input ~/en.wav
+  paddlespeech st --input ./en.wav
   ```
   Usage:
   ```bash
